+++ src/Client/SovereignClient/Program.cs
@@ -1,8 +1,8 @@
 /* Initialize SDL. */
 
 using System;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 using SDL2;
 using Serilog;
 using Sovereign.ClientCore;
@@ -14,18 +14,18 @@ using Sovereign.VeldridRenderer;
 var builder = Host.CreateApplicationBuilder(args);
 
 // Configure logging.
-var logger = new LoggerConfiguration()
+Log.Logger = new LoggerConfiguration()
     .ReadFrom.Configuration(builder.Configuration)
     .CreateLogger();
-builder.Services.AddLogging(
-    loggingBuilder => loggingBuilder.AddSerilog(dispose: true));
+builder.Logging.ClearProviders();
+builder.Logging.AddSerilog(dispose: true);
 
 // Initialize SDL.
 SDL.SDL_SetHint("SDL_WINDOWS_DPI_AWARENESS", "permonitorv2");
 var err = SDL.SDL_Init(SDL.SDL_INIT_EVERYTHING);
 if (err < 0)
 {
-    logger.Fatal("Error initializing SDL: {Error}", SDL.SDL_GetError());
+    Log.Logger.Fatal("Error initializing SDL: {Error}", SDL.SDL_GetError());
     Environment.Exit(1);
 }